backend/requirements.txt
# backend/requirements.txt
flask
openai
python-dotenv
web3
